Version 1
This version allows the user to select a browser, search engine and
search criteria using a python menu.  Once that is done, the application
will open a browser, go to the selected search engine and enter find the
search criteria.  It will then snap a picture of the browser displaying
the search results, count the number of images on the page and the
number of times the search criteria appears on the page.  This
information is then listed in a directory located in the same directory
the code is located.