Version 1.1
New menu options added but not functional.
a) The ability to count the images on any given page.
b) Validate Passwords

There is a new main menu in which these two options are added along with
the original functionality.  The new menu options are the first to
options on the main menu and the original functionality is the third
option.

I also created a function for exiting the application so that the
process is standardized making it easier to insert it into the new menu
functionality.