Version 2.0
This one has more functions.  Check the README.txt file for details.  An
extra menu was added at the beginning to facilitate adding the extra
functionality.  Also discovered that there is a bug currently in the way
that Chrome and Python work with Selenium.  I will need to do more
research in find a solution.